Struct ohsl::mesh2d::Mesh2D

pub struct Mesh2D<T> { /* fields omitted */ }

Implementations

impl<T: Clone + Number> Mesh2D<T>

pub fn new(x_nodes: Vector<f64>, y_nodes: Vector<f64>, nvars: usize) -> Self

Create a new 2D mesh

pub fn nvars(&self) -> usize

Return the number of variables stored at each node in the mesh

pub fn nnodes(&self) -> (usize, usize)

Return the number of nodal points in each direction

pub fn coord(&self, nodex: usize, nodey: usize) -> (f64, f64)

Return the spatial position of a specfied nodes as a tuple

pub fn xnodes(&self) -> Vector<f64>

Return the vector of x-nodal positions

pub fn ynodes(&self) -> Vector<f64>

Return the vector of y-nodal positions

pub fn set_nodes_vars(&mut self, nodex: usize, nodey: usize, vec: Vector<T>)

Set the variables stored at a specified node

pub fn get_nodes_vars(&self, nodex: usize, nodey: usize) -> Vector<T>

Get the vector of variables stored at a specified node

pub fn assign(&mut self, element: T)

Assign an element to all entries in the mesh

pub fn cross_section_xnode(&self, nodex: usize) -> Mesh1D<T, f64>

Return a cross section of the 2D mesh at a specified x node

pub fn cross_section_ynode(&self, nodey: usize) -> Mesh1D<T, f64>

Return a cross section of the 2D mesh at a specified y node

pub fn var_as_matrix(&self, var: usize) -> Matrix<T>

Return a matrix for a variable corresponding to each nodal point

pub fn apply(&mut self, func: &dyn Fn(f64, f64) -> T, var: usize)

Apply a function to the a specified variable in the mesh

impl Mesh2D<f64>

pub fn trapezium(&self, var: usize) -> f64

Integrate a given variable over the domain (trapezium rule)

pub fn square_trapezium(&self, var: usize) -> f64

Integrate the square of a given variable over the domain (trapezium rule)

impl<T: Display> Mesh2D<T>

pub fn output(&self, filename: &str, precision: usize)

Print the mesh to a file

pub fn output_var(&self, filename: &str, var: usize, precision: usize)

Print the mesh of a single variable to a file